fix(account): when display public keys from wallet prefix index and alias
